Fixed warnings associated with Network

+++ src/Engine/Network/Client.cpp
@@ -258,11 +258,11 @@ void Client::parseSnapshot(Packet& packet)
     }
 }
 
-int Client::receive(char* data)
+size_t Client::receive(char* data)
 {
     boost::system::error_code error;
 
-    int bytesReceived = m_Socket.receive_from(boost
+    size_t bytesReceived = m_Socket.receive_from(boost
         ::asio::buffer((void*)data, INPUTSIZE),
         m_ReceiverEndpoint,
         0, error);
@@ -390,7 +390,7 @@ void Client::identifyPacketLoss()
 bool Client::hasServerTimedOut()
 {
     // Time in ms
-    float timeSincePing = 1000 * (std::clock() - m_StartPingTime) / static_cast<double>(CLOCKS_PER_SEC);
+    double timeSincePing = 1000 * (std::clock() - m_StartPingTime) / static_cast<double>(CLOCKS_PER_SEC);
     if (timeSincePing > m_TimeoutMs) {
         // Clear everything and go to menu.
         LOG_INFO("Server has timed out, returning to menu, Beep Boop.");
